+++ src/main/java/ChatBotList.java
@@ -13,123 +13,13 @@ import java.time.format.DateTimeParseException;
  */
 public class ChatBotList {
     private ArrayList<Item> list;
-    
-
-    /**
-     * Encapsulates an item in the list
-     */
-    private static abstract class Item {
-        private String name;
-        private boolean isCompleted;
-
-        public Item(String name) {
-            this.name = name;
-            this.isCompleted = false;
-        }
-
-        /**
-         * Mark this item as completed.
-         */
-        public void markCompleted() {
-            this.isCompleted = true;
-        }
-
-        /**
-         * Removes the "completed" mark on this item.
-         */
-        public void unmarkCompleted() {
-            this.isCompleted = false;
-        }
-
-        @Override
-        public String toString() {
-            String rtnVal = "";
-            if (this.isCompleted) {
-                rtnVal += "[X] ";
-            } else {
-                rtnVal += "[ ] ";
-            }
-            return rtnVal + name;
-        }
-    }
-
-    /**
-     * Encapsulates an event with a start and end time/date.
-     */
-    private static class Event extends Item {
-        private LocalDateTime from;
-        private LocalDateTime to;
-
-        /**
-         * Creates a new event
-         * @param name Name of event.
-         * @param from Start time/date.
-         * @param to End time/date.
-         */
-        public Event(String name, LocalDateTime from, LocalDateTime to) {
-            super(name);
-            this.from = from;
-            this.to = to;
-        }
-
-        @Override
-        public String toString() {
-            String rtnVal = "";
-            if (super.isCompleted) {
-                rtnVal += "[E][X] ";
-            } else {
-                rtnVal += "[E][ ] ";
-            }
-            return rtnVal + super.name + " (from: " + DukeEnvironmentConstants.OUTPUT_FORMATTER.format(this.from) + " to: " + DukeEnvironmentConstants.OUTPUT_FORMATTER.format(this.to) + ")";
-        }
-    }
-    /**
-     * Encapsulates a deadline with a do-by time/date.
-     */
-    private static class Deadline extends Item {
-        private LocalDateTime by;
-
-        public Deadline(String input, LocalDateTime by) {
-            super(input);
-            this.by = by;
-        }
-
-        @Override
-        public String toString() {
-            String rtnVal = "";
-            if (super.isCompleted) {
-                rtnVal += "[D][X] ";
-            } else {
-                rtnVal += "[D][ ] ";
-            }
-            return rtnVal + super.name + " (by: " + DukeEnvironmentConstants.OUTPUT_FORMATTER.format(this.by) + ")";
-        }
-    }
-    /**
-     * Encapsulates a t0d0 task without time/date dependencies.
-     */
-    private static class Todo extends Item {
-        public Todo(String input) {
-            super(input);
-        }
-
-        @Override
-        public String toString() {
-            String rtnVal = "";
-            if (super.isCompleted) {
-                rtnVal += "[T][X] ";
-            } else {
-                rtnVal += "[T][ ] ";
-            }
-            return rtnVal + super.name;
-        }
-    }
-
 
     public ChatBotList() {
         this.list = new ArrayList<Item>();
     }
-
+    public ChatBotList(ArrayList<Item> list) {
+        this.list = list;
+    }
     public int getLength() {
         return this.list.size();
     }
@@ -144,15 +34,12 @@ public class ChatBotList {
         switch (type) {
             case EVENT:
                 this.list.add(new Event(queries[0], LocalDateTime.parse(queries[1],DukeEnvironmentConstants.FORMATTER1), LocalDateTime.parse(queries[2],DukeEnvironmentConstants.FORMATTER1)));
-                writeToSave();
                 break;
             case DEADLINE:
                 this.list.add(new Deadline(queries[0], LocalDateTime.parse(queries[1], DukeEnvironmentConstants.FORMATTER1)));
-                writeToSave();
                 break;
             case TODO:
                 this.list.add(new Todo(queries[0]));
-                writeToSave();
                 break;
             default:
                 break;
@@ -170,7 +57,6 @@ public class ChatBotList {
             throw new NotInChatBotListException();
         }
         this.list.get(index - 1).markCompleted();
-        writeToSave();
         return this.list.get(index - 1).toString();
     }
 
@@ -185,7 +71,6 @@ public class ChatBotList {
             throw new NotInChatBotListException();
         }
         this.list.get(index - 1).unmarkCompleted();
-        writeToSave();
         return this.list.get(index - 1).toString();
     }
 
@@ -201,77 +86,10 @@ public class ChatBotList {
         }
         Item itemToRemove = this.list.get(index - 1);
         this.list.remove(index - 1);
-        writeToSave();
         return itemToRemove.toString();
     }
 
-    /**
-     * Loads the list from the given file
-     */
-    public void loadList() {
-        this.list = new ArrayList<Item>();
-        try {
-            Scanner listReader = new Scanner(new File(DukeEnvironmentConstants.FILE_PATH));
-            while (listReader.hasNextLine()) {
-                String data = listReader.nextLine();
-                switch (data.substring(data.indexOf("["), data.indexOf("]") + 1)) {
-                    case "[T]":
-                        this.list.add(new Todo(data.substring(data.indexOf("[") + 7)));
-                        if (data.substring(data.indexOf("[") + 4, data.indexOf("[") + 5).equals("X")) {
-                            try {
-                                this.markItem(this.getLength());
-                            } catch (NotInChatBotListException e) {
-                                System.out.println(e.getMessage());
-                            } 
-                        }
-                        break;
-                    case "[E]":
-                        data.indexOf("(from: ");
-                        data.indexOf(" to: ");
-                        this.list.add(new Event(data.substring(data.indexOf("[") + 7, data.indexOf(" (from: ")), 
-                                LocalDateTime.parse(data.substring(data.indexOf("(from: ") + 7, data.indexOf(" to: ")), DukeEnvironmentConstants.OUTPUT_FORMATTER),
-                                LocalDateTime.parse(data.substring(data.indexOf("to: ") + 4, data.length() - 1), DukeEnvironmentConstants.OUTPUT_FORMATTER)));
-                        if (data.substring(data.indexOf("[") + 4, data.indexOf("[") + 5).equals("X")) {
-                            try {
-                                this.markItem(this.getLength());
-                            } catch (NotInChatBotListException e) {
-                                System.out.println(e.getMessage());
-                            } 
-                        }
-                        break;
-                    case "[D]":
-                        this.list.add(new Deadline(data.substring(data.indexOf("[") + 7, data.indexOf(" (by: ")),
-                                LocalDateTime.parse(data.substring(data.indexOf("(by: ") + 5, data.length()-1), DukeEnvironmentConstants.OUTPUT_FORMATTER)));
-                        if (data.substring(data.indexOf("[") + 4, data.indexOf("[") + 5).equals("X")) {
-                            try {
-                                this.markItem(this.getLength());
-                            } catch (NotInChatBotListException e) {
-                                System.out.println(e.getMessage());
-                            } 
-                        }
-                        break;
-                    default: throw new LoadListException();
-                }
-            }
-            listReader.close();
-        } catch (FileNotFoundException e) {
-            this.list = new ArrayList<Item>();
-        } catch (DateTimeParseException e) {
-            this.list = new ArrayList<Item>();
-        }catch (LoadListException e) {
-            this.list = new ArrayList<Item>();
-        } 
-    }
 
-    private void writeToSave() {
-        try {
-            BufferedWriter writer = new BufferedWriter(new FileWriter(DukeEnvironmentConstants.FILE_PATH));
-            writer.write(this.toString());
-            writer.close();
-        } catch (java.io.IOException e) {
-            System.out.println(e.getMessage());
-        }
-    }
     @Override
     public String toString() {
         String rtnVal = "";
