+++ src/main/java/Command.java
@@ -2,5 +2,5 @@
  * Encapsulates commands.
  */
 public abstract class Command {
-    public abstract void execute(ChatBotList list, Ui ui) throws ChatBotListException; 
+    public abstract void execute(ChatBotList list, Ui ui, Storage storage) throws ChatBotListException; 
 }
