+++ src/main/java/DeleteCommand.java
@@ -10,8 +10,9 @@ public class DeleteCommand extends Command{
     }
 
     @Override
-    public void execute(ChatBotList list, Ui ui) throws ChatBotListException{
+    public void execute(ChatBotList list, Ui ui, Storage storage) throws ChatBotListException{
         String res = list.deleteItem(this.idx);
         ui.print(String.format("Noted. I've removed this task:\n  " + res));
+        storage.writeToSave(list);
     }
 }
