+++ src/main/java/ExitCommand.java
@@ -1,5 +1,5 @@
 public class ExitCommand extends Command{
     @Override
-    public void execute(ChatBotList list, Ui ui) throws ChatBotListException{
+    public void execute(ChatBotList list, Ui ui, Storage storage) throws ChatBotListException{
     }
 }
