+++ src/main/java/MarkCommand.java
@@ -7,8 +7,9 @@ public class MarkCommand extends Command {
     }
 
     @Override
-    public void execute(ChatBotList list, Ui ui) throws ChatBotListException{
+    public void execute(ChatBotList list, Ui ui, Storage storage) throws ChatBotListException{
         String res = list.markItem(this.idx);
         ui.print(String.format("Nice! I've marked this task as done:\n  " + res));
+        storage.writeToSave(list);
     }
 }
