+++ src/main/java/UnmarkCommand.java
@@ -7,8 +7,9 @@ public class UnmarkCommand extends Command{
     }
 
     @Override
-    public void execute(ChatBotList list, Ui ui) throws ChatBotListException{
+    public void execute(ChatBotList list, Ui ui, Storage storage) throws ChatBotListException{
         String res = list.unmarkItem(this.idx);
         ui.print(String.format("OK, I've marked this task as not done yet:\n  " + res));
+        storage.writeToSave(list);
     }
 }
